requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
cssselect2==0.8.0
Flask==2.3.3
fonttools==4.59.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
reportlab==4.0.4
requests==2.32.4
stripe==6.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==2.3.7
zopfli==0.2.3.post1
playwright==1.48.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.0
